perf(habitat): split source-check rule modules
Replace the monolithic .habitat/source-check/source-rules.mjs policy with source-owned per-rule source-check modules plus a shared runtime. Load selected rule modules directly, scope Nx direct rule inputs to the owning module/runtime pair, and register .habitat as the Habitat artifact project so top-level rule artifacts have graph ownership without storing executable code in authored artifacts.

Validation:
- node --check tools/habitat-harness/src/domains/source-check/rule-runtime.mjs and all rule modules
- bun tools/habitat-harness/bin/dev.ts check --tool source-check --json
- bun tools/habitat-harness/bin/dev.ts check --owner @internal/habitat-harness --json
- bun run --cwd tools/habitat-harness check
- bun run --cwd tools/habitat-harness test -- test/rules/registry/contract.test.ts test/rules/registry/facts.test.ts test/service/hook-service.test.ts test/lib/rule-selection.test.ts
- bun run openspec -- validate deep-habitat-effect-source-check-rule-modules --strict
- bun run biome:ci
- git diff --check
