+++ tools/habitat-harness/src/domains/source-check/rules/adapter-base-standard-import.mjs
@@ -1,0 +1,12 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "adapter-base-standard-import";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime
+    .importRefs(file)
+    .filter((ref) => runtime.pathMatches(file, /^packages\/.*\.ts$/))
+    .filter(() => !file.path.includes("packages/civ7-adapter/"))
+    .filter((ref) => /\/base-standard\/.+/.test(ref.source))
+    .map((ref) => runtime.diagnostic(rule, file, ref.node));
+}
