+++ tools/habitat-harness/src/domains/source-check/rules/domain-deep-import.mjs
@@ -1,0 +1,12 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "domain-deep-import";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.sourceRefsMatching(
+    rule,
+    file,
+    /mods\/[^/]+\/src\/(?:recipes|maps)\/.*\.tsx?$/,
+    /^@mapgen\/domain\/[^/]+\/(?:ops\/.+|ops-by-id|rules\/.+|strategies\/+.+)$/
+  );
+}
