+++ tools/habitat-harness/src/domains/source-check/rules/domain-engine-imports.mjs
@@ -1,0 +1,14 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "domain-engine-imports";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime
+    .importRefs(file)
+    .filter((ref) => ref.kind === "import" && !ref.isTypeOnly)
+    .filter(() =>
+      runtime.pathMatches(file, /mods\/mod-swooper-maps\/src\/domain\/.*\/ops\/.*\.ts$/)
+    )
+    .filter((ref) => /(?:@swooper\/mapgen-core\/engine|@mapgen\/engine)$/.test(ref.source))
+    .map((ref) => runtime.diagnostic(rule, file, ref.node));
+}
