+++ tools/habitat-harness/src/domains/source-check/rules/habitat-adapter-domain-paths.mjs
@@ -1,0 +1,17 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "habitat-adapter-domain-paths";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.pathMatches(file, /tools\/habitat-harness\/src\/adapters\/grit\/.*\.ts$/) &&
+    /packages|apps\/|mods\/|mods\/mod-swooper-maps|apps\/mapgen-studio|\.civ7/.test(file.text)
+    ? [
+        runtime.diagnostic(
+          rule,
+          file,
+          undefined,
+          runtime.firstMatchingLine(file.text, /packages|apps\/|mods\/|\.civ7/)
+        ),
+      ]
+    : [];
+}
