+++ tools/habitat-harness/src/domains/source-check/rules/recipe-domain-surface.mjs
@@ -1,0 +1,16 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "recipe-domain-surface";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime
+    .importRefs(file)
+    .filter(() => runtime.pathMatches(file, /mods\/mod-swooper-maps\/src\/recipes\/.*\.ts$/))
+    .filter((ref) => /@mapgen\/domain\/[^/]+\/.+/.test(ref.source))
+    .filter((ref) => !/@mapgen\/domain\/[^/]+\/(?:ops|config\.js)$/.test(ref.source))
+    .filter(
+      (ref) =>
+        !/@mapgen\/domain\/[^/]+\/(?:ops\/.+|ops-by-id|rules\/.+|strategies\/.+)/.test(ref.source)
+    )
+    .map((ref) => runtime.diagnostic(rule, file, ref.node));
+}
