+++ tools/habitat-harness/src/domains/source-check/rules/recipe-runtime-domain-ops.mjs
@@ -1,0 +1,12 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "recipe-runtime-domain-ops";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.sourceRefsMatching(
+    rule,
+    file,
+    /mods\/[^/]+\/src\/recipes\/.*\/recipe\.ts$/,
+    /^@mapgen\/domain\/[^/]+$/
+  );
+}
