+++ tools/habitat-harness/src/domains/source-check/rules/relative-domain-imports.mjs
@@ -1,0 +1,7 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "relative-domain-imports";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.relativeDomainImportDiagnostics(rule, file);
+}
