+++ tools/habitat-harness/src/domains/source-check/rules/runtime-config-merge.mjs
@@ -1,0 +1,15 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "runtime-config-merge";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.pathMatches(
+    file,
+    /mods\/mod-swooper-maps\/src\/(?:recipes\/.*\/stages\/.*\/steps|domain\/.*\/ops)\/.*\.ts$/
+  )
+    ? [
+        ...runtime.emptyObjectNullish(file),
+        ...runtime.propertyCallExpressions(file, "Default", "Value"),
+      ].map((node) => runtime.diagnostic(rule, file, node))
+    : [];
+}
