+++ tools/habitat-harness/src/domains/source-check/rules/runtime-helper-redeclarations.mjs
@@ -1,0 +1,15 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "runtime-helper-redeclarations";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime
+    .helperRedeclarations(file)
+    .filter(() =>
+      runtime.pathMatches(
+        file,
+        /mods\/[^/]+\/src\/(?:recipes\/.*\/stages\/.*\/steps|domain\/.*\/ops\/.*\/strategies)\/.*\.ts$/
+      )
+    )
+    .map((node) => runtime.diagnostic(rule, file, node));
+}
