+++ tools/habitat-harness/src/domains/source-check/rules/runtime-run-validated.mjs
@@ -1,0 +1,15 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "runtime-run-validated";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.pathMatches(
+    file,
+    /mods\/[^/]+\/src\/(?:recipes\/.*\/stages\/.*\/steps|domain\/.*\/ops\/.*\/strategies)\/.*\.ts$/
+  )
+    ? [
+        ...runtime.callExpressions(file, "runValidated"),
+        ...runtime.propertyCallExpressions(file, "runValidated"),
+      ].map((node) => runtime.diagnostic(rule, file, node))
+    : [];
+}
