+++ tools/habitat-harness/src/domains/source-check/rules/runtime-validation-imports.mjs
@@ -1,0 +1,12 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "runtime-validation-imports";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.sourceRefsMatching(
+    rule,
+    file,
+    /mods\/[^/]+\/src\/(?:recipes\/.*\/stages\/.*\/steps|domain\/.*\/ops\/.*\/strategies)\/.*\.ts$/,
+    /^(?:@sinclair\/typebox\/value|@sinclair\/typebox\/compiler|@swooper\/mapgen-core\/compiler\/normalize|@swooper\/mapgen-core\/authoring\/validation|@swooper\/mapgen-core\/authoring\/op\/validation-surface)$/
+  );
+}
