+++ tools/habitat-harness/src/domains/source-check/rules/sdk-mapgen-entrypoint.mjs
@@ -1,0 +1,7 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "sdk-mapgen-entrypoint";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.sdkMapgenEntrypointDiagnostics(rule, file);
+}
