+++ tools/habitat-harness/src/domains/source-check/rules/sibling-stage-step-imports.mjs
@@ -1,0 +1,12 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "sibling-stage-step-imports";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.sourceRefsMatching(
+    rule,
+    file,
+    /mods\/mod-swooper-maps\/src\/recipes\/standard\/stages\/[^/]+\/.*\.ts$/,
+    /.*\.\.\/[^/]+\/steps\/.*/
+  );
+}
