+++ tools/habitat-harness/src/domains/source-check/rules/stage-contract-dependencies.mjs
@@ -1,0 +1,7 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "stage-contract-dependencies";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.stageContractDependencyDiagnostics(rule, file);
+}
