+++ tools/habitat-harness/src/domains/source-check/rules/step-contract-domain-surface.mjs
@@ -1,0 +1,14 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "step-contract-domain-surface";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime
+    .sourceRefsMatching(
+      rule,
+      file,
+      /mods\/[^/]+\/src\/recipes\/.*\/stages\/.*\/steps\/(?:.*\/)?(?:contract|[^/]+\.contract)\.ts$/,
+      /^@mapgen\/domain\/[^/]+\/.+$/
+    )
+    .filter(() => !/\/(?:__tests__|__type_tests__)\//.test(file.path));
+}
