+++ tools/habitat-harness/src/domains/source-check/rules/viz-contract-ownership.mjs
@@ -1,0 +1,7 @@
+import { sourceCheckRuntime as runtime } from "../rule-runtime.mjs";
+
+export const ruleId = "viz-contract-ownership";
+
+export function diagnosticsForRule(rule, file) {
+  return runtime.vizContractOwnershipDiagnostics(rule, file);
+}
