+++ tools/habitat-harness/src/plugin/target-definition-schema.ts
@@ -26,6 +26,7 @@ export const NxTargetDefinitionSchema = Type.Object(
 
 export const InferredProjectSchema = Type.Object(
   {
+    name: Type.Optional(Type.String({ minLength: 1 })),
     targets: Type.Record(Type.String({ minLength: 1 }), NxTargetDefinitionSchema),
   },
   { additionalProperties: false }
